// ehif-dbi/src/main/frontend/src/app/api/models/parish-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ParishionerSummaryDto } from '../models/parishioner-summary-dto';
import { PriestSummaryDto } from '../models/priest-summary-dto';
export interface ParishDto {
  foundedYear?: number;
  id?: number;
  location?: string;
  name?: string;
  parishioners?: Array<ParishionerSummaryDto>;
  priests?: Array<PriestSummaryDto>;
}